Change params of CopyFromTexture from Int -> Float

+++ Band/Source/Band/Public/BandTensor.h
@@ -24,6 +24,6 @@ public:
 
 	EBandStatus CopyFromBuffer(uint8* Buffer, int32 Bytes);
 	EBandStatus CopyFromBuffer(TArray<uint8> Buffer);
-	EBandStatus CopyFromTexture(UTexture2D* Texture, int32 Mean, int32 Std);
+	EBandStatus CopyFromTexture(UTexture2D* Texture, float Mean, float Std);
 	EBandStatus CopyToBuffer(TArray<uint8> Buffer);
 };
